Fix handling of dmon-ls service
Changed the way the parameters are assigned to an upstart job. Also
removed all the PID handling, because upstart does it all for us.
The restart action is indepotent, so it works both as start and
restart. The PID is then properly handled internally, therefore
we can set the value in the DMon to -1.